+++ src/html/types.ml
@@ -9,4 +9,4 @@ type toc = {
   children : toc list;
 }
 
-type breadcrumb = { href : string; name : string }
+type breadcrumb = { href : string; name : string; kind: Odoc_document.Url.Path.kind }